requirements.txt
discord.py[voice]>=2.0.0
yt-dlp>=2024.3.10
youtube-search-python>=1.6.6
PyNaCl>=1.5.0
python-dotenv>=1.0.0